fix dns and netmask parameters for onevnet
The parameter names opennebula expects are DNS and NETWORK_MASK not
dnsservers and netmask, so switch them when writing out the template
to be used when creating the onevnet. (Similar to vlanid to VLAN_ID
already in place.)

Also, the DNS value needs to output a space-separated string.

Also, fix string to array for ruby > 1.9 - when reading in dns from
onevnet list -x need to parse a space separated string. Ruby 1.8 had
a string.to_a but ruby >1.9 doesn't, so use .split(" ") instead (works
on both ruby versions).