parameter validation does not work as intended (tests)
parameter validation does not take place for parameters that are passed as parameters to the `one` class. This is a suit of tests that demonstrate the faulty behavior.

In addition, no check takes place that ssh keys are passed at all. We only check that a key that is passed is a string, the undef case remains unhandled.