reinstate old default behavior, add specs
by removing the default to $::ipaddress we reinstate the old default behavior of not configuring any onegate endpoint if no explicit parameter is passed. Passing the ip-address as default can now be done by passing the fact as parameter.

we also add a check that not both the onegate ip and the onegate endpoint are passed which would result in potentially confusing behavior.